+++ client/src/components/Auth/AuthContext.jsx
@@ -1,0 +1,26 @@
+import React, { createContext, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+export const AuthContext = createContext();
+
+
+const AuthProvider = ({ children }) => {
+    const [user, setUser] = useState(null);
+  
+    const login = (userData) => {
+      console.log(userData)
+      setUser(userData);
+    };
+  
+    const logout = () => {
+      setUser(null);
+    };
+  
+    return (
+      <AuthContext.Provider value={{ user, login, logout }}>
+        {children}
+      </AuthContext.Provider>
+    );
+  };
+
+export default AuthProvider;
