+++ client/src/index.js
@@ -1,11 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { GoogleOAuthProvider } from '@react-oauth/google';
+import { BrowserRouter as Router } from 'react-router-dom';
+import AuthProvider from './components/Auth/AuthContext';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+      <AuthProvider>
+        <Router>
+          <App />
+        </Router>
+      </AuthProvider>
   </React.StrictMode>
 );
