+++ client/src/screens/Login.jsx
@@ -1,0 +1,37 @@
+import React, {useContext} from 'react';
+import { GoogleOAuthProvider, GoogleLogin } from '@react-oauth/google';
+import { LoginSocialFacebook } from "reactjs-social-login";
+import { FacebookLoginButton } from "react-social-login-buttons";
+import { AuthContext } from '../components/Auth/AuthContext';
+
+
+const Login = () => {
+    const responseFacebook = (response) => {
+        console.log(response);
+      }
+
+    const {user, login, logout} = useContext(AuthContext);
+      
+    return (
+      <div>
+      <LoginSocialFacebook
+      appId="464555689517211"
+      onResolve={login}
+      onReject={(error) => {
+        console.log(error);
+      }}
+      >
+        <FacebookLoginButton />
+      </LoginSocialFacebook>
+      <GoogleOAuthProvider clientId="174570537860-6vkd39p098a9mnb66kpi9tlistc3i224.apps.googleusercontent.com">
+
+          <GoogleLogin
+          onSuccess={responseFacebook}
+        />
+      </GoogleOAuthProvider>
+      </div>
+
+    );
+};
+
+export default Login;
